feat: change diretory name
