--- 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/274-follow-material/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "274-follow-material",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "mysql": "^2.18.1",
-    "redis": "^3.1.2",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
-  },
-  "devDependencies": {
-    "@types/mysql": "^2.15.21",
-    "@types/redis": "^4.0.11"
-  }
-}
