--- 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/274-follow-material/src/Connectors/index.ts
@@ -1,11 +1,0 @@
-import Connector, { ConnectorConstructor } from './Connector';
-import MySQLConnector from './mysql/MySQLConnector';
-import RedisConnector from './redis/RedisConnector';
-
-export default Connector;
-export {
-  Connector,
-  ConnectorConstructor,
-  MySQLConnector,
-  RedisConnector,
-};
