--- 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/274-follow-material/src/index.ts
@@ -1,39 +1,0 @@
-// ./src/index.ts
-import { Connector, RedisConnector } from "./Connectors";
-
-const token = 'ce42033d-9133-457a-a1a1-41ac0b63a333';
-
-const conn = new RedisConnector({
-  host: 'redisdb',
-  port: 6379});
-
-const main = (connector: Connector) => {
-  connector.firstCount(token);
-
-  const logCount = async () => {
-    try {
-      const count = await connector.getCount(token);
-      console.log(count);
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  const doSomeIncrements = () => {
-    logCount();
-    connector.incrementCount(token);
-    connector.incrementCount(token);
-    connector.incrementCount(token);
-  };
-
-  const incrementsInterval = setInterval(doSomeIncrements, 500);
-
-  setTimeout(() => {
-    clearInterval(incrementsInterval);
-    logCount();
-    connector.clearCount(token);
-    connector.closeConnection();
-  }, 5500);
-};
-
-main(conn);
