+++ 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/275-follow-material/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "solid-2",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "start": "npx ts-node src"
+  },
+  "license": "ISC",
+  "devDependencies": {
+    "@types/mysql": "^2.15.19",
+    "@types/redis": "^2.8.32"
+  },
+  "dependencies": {
+    "ts-node": "^10.2.1",
+    "mysql": "^2.18.1",
+    "redis": "^3.1.2",
+    "typescript": "^4.4.3"
+  }
+}
