+++ 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/275-follow-material/src/Connectors/Connector.ts
@@ -1,11 +1,11 @@
-import { ConnectionTimeoutError } from "redis";
-
-export default interface Connector {
-  getCount(token: string): number | Promise<number>;
-
-  incrementCount(token: string): void;
+export interface ReadOnlyConnector {
+  getCount(token: string): number | Promise <number>;
 
   closeConnection(): void;
+}
+
+export default interface Connector {
+  incrementCount(token: string): void;
 
   clearCount(token: string): void;
