+++ 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/275-follow-material/src/Connectors/index.ts
@@ -1,0 +1,16 @@
+// ./src/Connectors/index.ts
+
+import Connector, { ReadOnlyConnector, ConnectorConstructor } from "./Connector";
+import MySQLConnector from "./mysql/MySQLConnector";
+import ReadOnlyRedisConnector from "./redis/ReadOnlyRedisConnector";
+import RedisConnector from "./redis/RedisConnector";
+
+export default Connector;
+export {
+  Connector,
+  ConnectorConstructor,
+  MySQLConnector,
+  RedisConnector,
+  ReadOnlyRedisConnector,
+  ReadOnlyConnector
+};
