+++ 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/275-follow-material/src/Connectors/mysql/MySQLConnector.ts
