+++ 03-back-end/bloco-27-programacao-orienta-a-objetos-poo-e-solid/dia-05-principios-l-e-i/275-Exercises/275-follow-material/src/Connectors/redis/RedisConnector.ts
@@ -1,0 +1,17 @@
+// ./src/Connectors/redis/RedisConnector.ts
+
+import Connector from '../Connector';
+import ReadOnlyRedisConnector from './ReadOnlyRedisConnector';
+
+export default class RedisConnector extends ReadOnlyRedisConnector implements Connector {
+
+  public incrementCount(token: string): void {
+    this.client.incr(token);
+  }
+
+  public clearCount(token: string): void {
+    this.client.set(token, '0');
+  }
+
+  public firstCount = this.clearCount;
+}
